requirements.txt
fastapi>=0.95.0
uvicorn>=0.15.0
langchain>=0.0.267
langchain-deepseek>=0.0.1
python-dotenv>=0.19.0
pydantic>=2.0.0
httpx>=0.23.0
pytest>=7.0.0
websockets>=10.0 